tests: move unittest from OWDifferentialExpression to test folder
